FEATURE: added local containerisation
