+++ .air.toml
@@ -1,0 +1,37 @@
+root = "."
+testdata_dir = "testdata"
+tmp_dir = "air-tmp"
+
+[build]
+  args_bin = []
+  bin = "./air-tmp/tiger-payment"
+  cmd = "go build -o ./air-tmp/tiger-payment ."
+  delay = 1000
+  exclude_dir = ["assets", "air-tmp", "vendor", "testdata"]
+  exclude_file = []
+  exclude_regex = ["_test.go"]
+  exclude_unchanged = false
+  follow_symlink = false
+  full_bin = "./air-tmp/tiger-payment DEV tiger-db tigerguy tigercoders tigerlily2 disable 5432"
+  include_dir = []
+  include_ext = ["go", "tpl", "tmpl", "html"]
+  kill_delay = "0s"
+  log = "build-errors.log"
+  send_interrupt = false
+  stop_on_error = true
+
+[color]
+  app = ""
+  build = "yellow"
+  main = "magenta"
+  runner = "green"
+  watcher = "cyan"
+
+[log]
+  time = false
+
+[misc]
+  clean_on_exit = false
+
+[screen]
+  clear_on_rebuild = false
